상품상세 조회 추가

+++ src/main/java/com/meongnyang/shop/dto/request/admin/ReqModifyProductDto.java
@@ -46,7 +46,7 @@ public class ReqModifyProductDto {
     private Long arrivalQuantity;
     private Long minAlertQuantity;
     private Long alertSetting;
-    private Long outOfStock;
+    private int outOfStock;
 
     public Product toEntity() {
         return Product.builder()
